# swagger_server/models/product_review.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from database.dbmodels.product_review_dbmodel import ProductReviewDbModel
from swagger_server.models.base_model_ import Model
from swagger_server import util


class ProductReview(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, review_id: str = None, product_id: str = None, author_id: str = None, rating: int = None,
                 description: str = None):  # noqa: E501
        """ProductReview - a model defined in Swagger

        :param review_id: The review_id of this ProductReview.  # noqa: E501
        :type review_id: str
        :param author_id: The author_id of this ProductReview.  # noqa: E501
        :type author_id: str
        :param rating: The rating of this ProductReview.  # noqa: E501
        :type rating: int
        :param description: The description of this ProductReview.  # noqa: E501
        :type description: str
        """
        self.swagger_types = {
            'review_id': str,
            'product_id': str,
            'author_id': str,
            'rating': int,
            'description': str
        }

        self.attribute_map = {
            'review_id': 'review_id',
            'product_id': 'product_id',
            'author_id': 'author_id',
            'rating': 'rating',
            'description': 'description'
        }
        self._review_id = review_id
        self._product_id = product_id
        self._author_id = author_id
        self._rating = rating
        self._description = description

    @classmethod
    def from_dict(cls, dikt) -> 'ProductReview':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The ProductReview of this ProductReview.  # noqa: E501
        :rtype: ProductReview
        """
        return util.deserialize_model(dikt, cls)

    @property
    def review_id(self) -> str:
        """Gets the review_id of this ProductReview.


        :return: The review_id of this ProductReview.
        :rtype: str
        """
        return self._review_id

    @review_id.setter
    def review_id(self, review_id: str):
        """Sets the review_id of this ProductReview.


        :param review_id: The review_id of this ProductReview.
        :type review_id: str
        """

        self._review_id = review_id

    @property
    def product_id(self) -> str:
        """Gets the product_id of this ProductReview.


        :return: The product_id of this ProductReview.
        :rtype: str
        """
        return self._product_id

    @product_id.setter
    def product_id(self, product_id: str):
        """Sets the product_id of this ProductReview.


        :param product_id: The product_id of this ProductReview.
        :type product_id: str
        """
        if product_id is None:
            raise ValueError("Invalid value for `product_id`, must not be `None`")  # noqa: E501

        self._product_id = product_id

    @property
    def author_id(self) -> str:
        """Gets the author_id of this ProductReview.


        :return: The author_id of this ProductReview.
        :rtype: str
        """
        return self._author_id

    @author_id.setter
    def author_id(self, author_id: str):
        """Sets the author_id of this ProductReview.


        :param author_id: The author_id of this ProductReview.
        :type author_id: str
        """
        if author_id is None:
            raise ValueError("Invalid value for `author_id`, must not be `None`")  # noqa: E501

        self._author_id = author_id

    @property
    def rating(self) -> int:
        """Gets the rating of this ProductReview.


        :return: The rating of this ProductReview.
        :rtype: int
        """
        return self._rating

    @rating.setter
    def rating(self, rating: int):
        """Sets the rating of this ProductReview.


        :param rating: The rating of this ProductReview.
        :type rating: int
        """
        allowed_values = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]  # noqa: E501
        if rating not in allowed_values:
            raise ValueError(
                "Invalid value for `rating` ({0}), must be one of {1}"
                .format(rating, allowed_values)
            )

        self._rating = rating

    @property
    def description(self) -> str:
        """Gets the description of this ProductReview.


        :return: The description of this ProductReview.
        :rtype: str
        """
        return self._description

    @description.setter
    def description(self, description: str):
        """Sets the description of this ProductReview.


        :param description: The description of this ProductReview.
        :type description: str
        """
        if description is None:
            raise ValueError("Invalid value for `description`, must not be `None`")  # noqa: E501

        self._description = description

    def toProductReviewDbModel(self):
        product_review_dbmodel = ProductReviewDbModel()
        product_review_dbmodel.product_id = self.product_id
        product_review_dbmodel.author_id = self.author_id
        product_review_dbmodel.rating = self.rating
        product_review_dbmodel.description = self.description
        return product_review_dbmodel

    def fromProductReviewDbModel(self, product_review_dbmodel):
        self.review_id = str(product_review_dbmodel.id)
        self.product_id = product_review_dbmodel.product_id
        self.author_id = product_review_dbmodel.author_id
        self.rating = product_review_dbmodel.rating
        self.description = product_review_dbmodel.description