# swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Bakers hub OpenAPI spec
  version: 0.0.1
servers:
- url: http://localhost:8080/v1
  description: Dev server
paths:
...
  /products:
    get:
      tags:
      - Products
      summary: Returns products of all bakers
      operationId: get_products
      parameters:
      - name: title
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
        example: cake
      - name: baker_id
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
        example: "1"
      - name: rating
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: number
        example: 6.7
      responses:
        "200":
          description: Successful response with list of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
                x-content-type: application/json
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultError'
      x-openapi-router-controller: swagger_server.controllers.products_controller
    post:
      tags:
      - Products
      summary: Creates a new product
      operationId: create_product
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
        required: true
      responses:
        "200":
          description: Successful response with new product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultError'
      x-openapi-router-controller: swagger_server.controllers.products_controller
  /products/{product_id}:
    get:
      tags:
      - Products
      summary: Returns product with requested ID
      operationId: get_product_by_id
      parameters:
      - name: product_id
        in: path
        description: Product's ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example: "1"
      responses:
        "200":
          description: Successful response with relevant product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultError'
      x-openapi-router-controller: swagger_server.controllers.products_controller
    put:
      tags:
      - Products
      summary: Updating existing product
      operationId: update_product_by_id
      parameters:
      - name: product_id
        in: path
        description: Product's ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example: "1"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
        required: true
      responses:
        "200":
          description: Successful response with updated product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultError'
      x-openapi-router-controller: swagger_server.controllers.products_controller
    delete:
      tags:
      - Products
      description: Deletes product with requested Id
      operationId: delete_product_by_id
      parameters:
      - name: product_id
        in: path
        description: Product's ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example: "1"
      responses:
        "200":
          description: Successfully deleted
          content:
            application/json: {}
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultError'
      x-openapi-router-controller: swagger_server.controllers.products_controller
  /products/{product_id}/reviews:
    get:
      tags:
      - Products
      summary: Returns product reviews
      operationId: get_product_reviews_by_product_id
      parameters:
      - name: product_id
        in: path
        description: Product's ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example: "1"
      responses:
        "200":
          description: Successful response with list of product reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductReview'
                x-content-type: application/json
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultError'
      x-openapi-router-controller: swagger_server.controllers.products_controller
    post:
      tags:
      - Products
      summary: Creates product review
      operationId: create_product_review
      parameters:
      - name: product_id
        in: path
        description: Product's ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example: "1"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductReview'
        required: true
      responses:
        "200":
          description: Successful response with product review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductReview'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultError'
      x-openapi-router-controller: swagger_server.controllers.products_controller
    delete:
      tags:
      - Products
      summary: Deletes product review
      operationId: delete_product_review_by_id
      parameters:
      - name: product_id
        in: path
        description: Product's ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example: "1"
      - name: review_id
        in: query
        description: Review's ID
        required: true
        style: form
        explode: true
        schema:
          type: string
          example: "1"
      responses:
        "200":
          description: Successfully deleted
          content:
            application/json: {}
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultError'
      x-openapi-router-controller: swagger_server.controllers.products_controller
...
components:
  schemas:
...
    Product:
      required:
      - baker_id
      - description
      - execution
      - ingredients
      - price
      - status
      - title
      type: object
      properties:
        product_id:
          type: string
          example: "1"
        baker_id:
          type: string
          example: "1"
        title:
          type: string
          example: cake
        price:
          type: string
          example: 1000/kg
        description:
          type: string
          example: diet cake with no sugar
        ingredients:
          type: string
          example: "flour, eggs, rotten tomatoes"
        execution:
          type: string
          example: 1 day
        status:
          type: string
          enum:
          - available
          - not available
        rating:
          type: number
          example: 7
        reviews_quantity:
          type: integer
          format: int64
          example: 1
      example:
        reviews_quantity: 1
        execution: 1 day
        price: 1000/kg
        product_id: "1"
        rating: 7
        description: diet cake with no sugar
        ingredients: "flour, eggs, rotten tomatoes"
        baker_id: "1"
        title: cake
        status: available
...
    ProductReview:
      required:
      - author_id
      - product_id
      - description
      - rating
      type: object
      properties:
        review_id:
          type: string
          example: "1"
        author_id:
          type: string
          example: "1"
        product_id:
          type: string
          example: "1"
        rating:
          type: integer
          example: 9
          enum:
          - 1
          - 2
          - 3
          - 4
          - 5
          - 6
          - 7
          - 8
          - 9
          - 10
        description:
          type: string
          example: very good very well
      example:
        review_id: "1"
        rating: 9
        description: very good very well
        author_id: "1"
    DefaultError:
      required:
      - code
      - message
      type: object
      properties:
        code:
          type: integer
        message:
          type: string